+++ src/utils/scheduler.py
@@ -1,16 +1,16 @@
 from apscheduler.schedulers.background import BackgroundScheduler
 from apscheduler.triggers.interval import IntervalTrigger
-from src.tasks.policies import task_run_policies
+from src.tasks.policies import run_policies
 import atexit
 
 
-def schedule():
+def schedule(trigger=IntervalTrigger(hours=2)):
     scheduler = BackgroundScheduler()
     scheduler.start()
     scheduler.add_job(
-        func=task_run_policies,
-        trigger=IntervalTrigger(minutes=20),
-        name='Schedule task "cleanup time-series" every 20 minutes',
-        replace_existing=True
+        func=run_policies,
+        trigger=trigger,
+        replace_existing=True,
+        name="Clean-up Prometheus time-series"
     )
     atexit.register(lambda: scheduler.shutdown())